requirements.txt
fastapi
uvicorn[standard]
tavily-python
python-dotenv